Extract TabPageControl nested classes
