+++ src/System.Windows.Forms/src/System/Windows/Forms/TabControl.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Collections;
 using System.Collections.Specialized;
 using System.ComponentModel;
 using System.Diagnostics;
@@ -2192,540 +2191,5 @@ namespace System.Windows.Forms
         private bool GetState(State state) => _tabControlState[(int)state];
 
         private void SetState(State state, bool value) => _tabControlState[(int)state] = value;
-
-        public class TabPageCollection : IList
-        {
-            private readonly TabControl owner;
-            ///  A caching mechanism for key accessor
-            ///  We use an index here rather than control so that we don't have lifetime
-            ///  issues by holding on to extra references.
-            private int lastAccessedIndex = -1;
-
-            public TabPageCollection(TabControl owner)
-            {
-                this.owner = owner ?? throw new ArgumentNullException(nameof(owner));
-            }
-
-            public virtual TabPage this[int index]
-            {
-                get
-                {
-                    return owner.GetTabPage(index);
-                }
-                set
-                {
-                    owner.SetTabPage(index, value, value.GetTCITEM());
-                }
-            }
-
-            object IList.this[int index]
-            {
-                get
-                {
-                    return this[index];
-                }
-                set
-                {
-                    if (value is TabPage)
-                    {
-                        this[index] = (TabPage)value;
-                    }
-                    else
-                    {
-                        throw new ArgumentException(nameof(value));
-                    }
-                }
-            }
-
-            /// <summary>
-            ///  Retrieves the child control with the specified key.
-            /// </summary>
-            public virtual TabPage this[string key]
-            {
-                get
-                {
-                    // We do not support null and empty string as valid keys.
-                    if (string.IsNullOrEmpty(key))
-                    {
-                        return null;
-                    }
-
-                    // Search for the key in our collection
-                    int index = IndexOfKey(key);
-                    if (IsValidIndex(index))
-                    {
-                        return this[index];
-                    }
-                    else
-                    {
-                        return null;
-                    }
-
-                }
-            }
-
-            [Browsable(false)]
-            public int Count
-            {
-                get
-                {
-                    return owner._tabPageCount;
-                }
-            }
-
-            object ICollection.SyncRoot
-            {
-                get
-                {
-                    return this;
-                }
-            }
-
-            bool ICollection.IsSynchronized
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            bool IList.IsFixedSize
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            public bool IsReadOnly
-            {
-                get
-                {
-                    return false;
-                }
-            }
-
-            public void Add(TabPage value)
-            {
-
-                if (value == null)
-                {
-                    throw new ArgumentNullException(nameof(value));
-                }
-
-                owner.Controls.Add(value);
-            }
-
-            int IList.Add(object value)
-            {
-                if (value is TabPage)
-                {
-                    Add((TabPage)value);
-                    return IndexOf((TabPage)value);
-                }
-                else
-                {
-                    throw new ArgumentException(nameof(value));
-                }
-            }
-
-            // <-- NEW ADD OVERLOADS FOR WHIDBEY
-
-            public void Add(string text)
-            {
-                TabPage page = new TabPage
-                {
-                    Text = text
-                };
-                Add(page);
-            }
-
-            public void Add(string key, string text)
-            {
-                TabPage page = new TabPage
-                {
-                    Name = key,
-                    Text = text
-                };
-                Add(page);
-            }
-
-            public void Add(string key, string text, int imageIndex)
-            {
-                TabPage page = new TabPage
-                {
-                    Name = key,
-                    Text = text,
-                    ImageIndex = imageIndex
-                };
-                Add(page);
-            }
-
-            public void Add(string key, string text, string imageKey)
-            {
-                TabPage page = new TabPage
-                {
-                    Name = key,
-                    Text = text,
-                    ImageKey = imageKey
-                };
-                Add(page);
-            }
-
-            // END - NEW ADD OVERLOADS FOR WHIDBEY -->
-
-            public void AddRange(TabPage[] pages)
-            {
-                if (pages == null)
-                {
-                    throw new ArgumentNullException(nameof(pages));
-                }
-                foreach (TabPage page in pages)
-                {
-                    Add(page);
-                }
-            }
-
-            public bool Contains(TabPage page)
-            {
-
-                //check for the page not to be null
-                if (page == null)
-                {
-                    throw new ArgumentNullException(nameof(page));
-                }
-                //end check
-
-                return IndexOf(page) != -1;
-            }
-
-            bool IList.Contains(object page)
-            {
-                if (page is TabPage)
-                {
-                    return Contains((TabPage)page);
-                }
-                else
-                {
-                    return false;
-                }
-            }
-
-            /// <summary>
-            ///  Returns true if the collection contains an item with the specified key, false otherwise.
-            /// </summary>
-            public virtual bool ContainsKey(string key)
-            {
-                return IsValidIndex(IndexOfKey(key));
-            }
-
-            public int IndexOf(TabPage page)
-            {
-
-                //check for the page not to be null
-                if (page == null)
-                {
-                    throw new ArgumentNullException(nameof(page));
-                }
-                //end check
-
-                for (int index = 0; index < Count; ++index)
-                {
-                    if (this[index] == page)
-                    {
-                        return index;
-                    }
-                }
-                return -1;
-            }
-
-            int IList.IndexOf(object page)
-            {
-                if (page is TabPage)
-                {
-                    return IndexOf((TabPage)page);
-                }
-                else
-                {
-                    return -1;
-                }
-            }
-            /// <summary>
-            ///  The zero-based index of the first occurrence of value within the entire CollectionBase, if found; otherwise, -1.
-            /// </summary>
-            public virtual int IndexOfKey(string key)
-            {
-                // Step 0 - Arg validation
-                if (string.IsNullOrEmpty(key))
-                {
-                    return -1; // we dont support empty or null keys.
-                }
-
-                // step 1 - check the last cached item
-                if (IsValidIndex(lastAccessedIndex))
-                {
-                    if (WindowsFormsUtils.SafeCompareStrings(this[lastAccessedIndex].Name, key, /* ignoreCase = */ true))
-                    {
-                        return lastAccessedIndex;
-                    }
-                }
-
-                // step 2 - search for the item
-                for (int i = 0; i < Count; i++)
-                {
-                    if (WindowsFormsUtils.SafeCompareStrings(this[i].Name, key, /* ignoreCase = */ true))
-                    {
-                        lastAccessedIndex = i;
-                        return i;
-                    }
-                }
-
-                // step 3 - we didn't find it.  Invalidate the last accessed index and return -1.
-                lastAccessedIndex = -1;
-                return -1;
-            }
-
-            /// <summary>
-            ///  Inserts the supplied Tabpage at the given index.
-            /// </summary>
-            public void Insert(int index, TabPage tabPage)
-            {
-                owner.InsertItem(index, tabPage);
-                try
-                {
-                    // See InsertingItem property
-                    owner.InsertingItem = true;
-                    owner.Controls.Add(tabPage);
-                }
-                finally
-                {
-                    owner.InsertingItem = false;
-                }
-                owner.Controls.SetChildIndex(tabPage, index);
-            }
-
-            void IList.Insert(int index, object tabPage)
-            {
-                if (tabPage is TabPage)
-                {
-                    Insert(index, (TabPage)tabPage);
-                }
-                else
-                {
-                    throw new ArgumentException(nameof(tabPage));
-                }
-            }
-
-            // <-- NEW INSERT OVERLOADS FOR WHIDBEY
-
-            public void Insert(int index, string text)
-            {
-                TabPage page = new TabPage
-                {
-                    Text = text
-                };
-                Insert(index, page);
-            }
-
-            public void Insert(int index, string key, string text)
-            {
-                TabPage page = new TabPage
-                {
-                    Name = key,
-                    Text = text
-                };
-                Insert(index, page);
-            }
-
-            public void Insert(int index, string key, string text, int imageIndex)
-            {
-                TabPage page = new TabPage
-                {
-                    Name = key,
-                    Text = text
-                };
-                Insert(index, page);
-                // ImageKey and ImageIndex require parenting...
-                page.ImageIndex = imageIndex;
-            }
-
-            public void Insert(int index, string key, string text, string imageKey)
-            {
-                TabPage page = new TabPage
-                {
-                    Name = key,
-                    Text = text
-                };
-                Insert(index, page);
-                // ImageKey and ImageIndex require parenting...
-                page.ImageKey = imageKey;
-            }
-
-            // END - NEW INSERT OVERLOADS FOR WHIDBEY -->
-
-            /// <summary>
-            ///  Determines if the index is valid for the collection.
-            /// </summary>
-            private bool IsValidIndex(int index)
-            {
-                return ((index >= 0) && (index < Count));
-            }
-
-            public virtual void Clear()
-            {
-                owner.RemoveAll();
-            }
-
-            void ICollection.CopyTo(Array dest, int index)
-            {
-                if (Count > 0)
-                {
-                    System.Array.Copy(owner.GetTabPages(), 0, dest, index, Count);
-                }
-            }
-
-            public IEnumerator GetEnumerator()
-            {
-                TabPage[] tabPages = owner.GetTabPages();
-                if (tabPages != null)
-                {
-                    return tabPages.GetEnumerator();
-                }
-                else
-                {
-                    return Array.Empty<TabPage>().GetEnumerator();
-                }
-            }
-
-            public void Remove(TabPage value)
-            {
-
-                //check for the value not to be null
-                if (value == null)
-                {
-                    throw new ArgumentNullException(nameof(value));
-                }
-                //end check
-                owner.Controls.Remove(value);
-            }
-
-            void IList.Remove(object value)
-            {
-                if (value is TabPage)
-                {
-                    Remove((TabPage)value);
-                }
-            }
-
-            public void RemoveAt(int index)
-            {
-                owner.Controls.RemoveAt(index);
-            }
-
-            /// <summary>
-            ///  Removes the child control with the specified key.
-            /// </summary>
-            public virtual void RemoveByKey(string key)
-            {
-                int index = IndexOfKey(key);
-                if (IsValidIndex(index))
-                {
-                    RemoveAt(index);
-                }
-            }
-
-        }
-
-        /// <summary>
-        ///  Collection of controls...
-        /// </summary>
-        [ComVisible(false)]
-        public new class ControlCollection : Control.ControlCollection
-        {
-            private readonly TabControl owner;
-
-            /*C#r: protected*/
-
-            public ControlCollection(TabControl owner)
-            : base(owner)
-            {
-                this.owner = owner;
-            }
-
-            public override void Add(Control value)
-            {
-                if (!(value is TabPage))
-                {
-                    throw new ArgumentException(string.Format(SR.TabControlInvalidTabPageType, value.GetType().Name));
-                }
-
-                TabPage tabPage = (TabPage)value;
-
-                // See InsertingItem property
-                if (!owner.InsertingItem)
-                {
-                    if (owner.IsHandleCreated)
-                    {
-                        owner.AddTabPage(tabPage, tabPage.GetTCITEM());
-                    }
-                    else
-                    {
-                        owner.Insert(owner.TabCount, tabPage);
-                    }
-                }
-
-                base.Add(tabPage);
-                tabPage.Visible = false;
-
-                // Without this check, we force handle creation on the tabcontrol
-                // which is not good at all of there are any OCXs on it.
-                //
-                if (owner.IsHandleCreated)
-                {
-                    tabPage.Bounds = owner.DisplayRectangle;
-                }
-
-                // site the tabPage if necessary.
-                ISite site = owner.Site;
-                if (site != null)
-                {
-                    ISite siteTab = tabPage.Site;
-                    if (siteTab == null)
-                    {
-                        IContainer container = site.Container;
-                        if (container != null)
-                        {
-                            container.Add(tabPage);
-                        }
-                    }
-                }
-                owner.ApplyItemSize();
-                owner.UpdateTabSelection(false);
-
-            }
-
-            public override void Remove(Control value)
-            {
-                base.Remove(value);
-                if (!(value is TabPage))
-                {
-                    return;
-                }
-                int index = owner.FindTabPage((TabPage)value);
-                int curSelectedIndex = owner.SelectedIndex;
-
-                if (index != -1)
-                {
-                    owner.RemoveTabPage(index);
-                    if (index == curSelectedIndex)
-                    {
-                        owner.SelectedIndex = 0; //Always select the first tabPage is the Selected TabPage is removed.
-                    }
-                }
-                owner.UpdateTabSelection(false);
-            }
-
-        }
-
     }
 }
